requirements.txt
asgiref==3.8.0
backports.zoneinfo==0.2.1
Django==4.2.11
django-filter==24.1
djangorestframework==3.15.0
importlib-metadata==7.1.0
Markdown==3.6
pillow==10.2.0
sqlparse==0.4.4
typing-extensions==4.10.0
zipp==3.18.1
